+++ enigma/src/main/java/org/quiltmc/enigma/api/translation/mapping/serde/MappingFormat.java
@@ -1,44 +1,15 @@
 package org.quiltmc.enigma.api.translation.mapping.serde;
 
-import com.google.common.io.MoreFiles;
 import org.quiltmc.enigma.api.ProgressListener;
 import org.quiltmc.enigma.api.translation.mapping.EntryMapping;
 import org.quiltmc.enigma.api.translation.mapping.MappingDelta;
-import org.quiltmc.enigma.api.translation.mapping.serde.enigma.EnigmaMappingsReader;
-import org.quiltmc.enigma.api.translation.mapping.serde.enigma.EnigmaMappingsWriter;
-import org.quiltmc.enigma.api.translation.mapping.serde.proguard.ProguardMappingsReader;
-import org.quiltmc.enigma.api.translation.mapping.serde.srg.SrgMappingsWriter;
-import org.quiltmc.enigma.api.translation.mapping.serde.tinyv2.TinyV2Reader;
-import org.quiltmc.enigma.api.translation.mapping.serde.tinyv2.TinyV2Writer;
 import org.quiltmc.enigma.api.translation.mapping.tree.EntryTree;
-import org.quiltmc.enigma.impl.translation.FileType;
 
 import javax.annotation.Nullable;
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.Objects;
 
-public enum MappingFormat {
-	ENIGMA_FILE(EnigmaMappingsWriter.FILE, EnigmaMappingsReader.FILE, FileType.ENIGMA_MAPPING),
-	ENIGMA_DIRECTORY(EnigmaMappingsWriter.DIRECTORY, EnigmaMappingsReader.DIRECTORY, FileType.ENIGMA_DIRECTORY),
-	ENIGMA_ZIP(EnigmaMappingsWriter.ZIP, EnigmaMappingsReader.ZIP, FileType.ENIGMA_ZIP),
-	TINY_V2(new TinyV2Writer("intermediary", "named"), new TinyV2Reader(), FileType.TINY_V2),
-	SRG_FILE(SrgMappingsWriter.INSTANCE, null, FileType.SRG),
-	PROGUARD(null, ProguardMappingsReader.INSTANCE, FileType.PROGUARD);
-
-	private final MappingsWriter writer;
-	private final MappingsReader reader;
-	private final FileType fileType;
-
-	MappingFormat(MappingsWriter writer, MappingsReader reader, FileType fileType) {
-		this.writer = writer;
-		this.reader = reader;
-		this.fileType = fileType;
-	}
-
+public class MappingFormat {
 	public void write(EntryTree<EntryMapping> mappings, Path path, MappingSaveParameters saveParameters) {
 		this.write(mappings, MappingDelta.added(mappings), path, ProgressListener.createEmpty(), saveParameters);
 	}
@@ -84,59 +55,5 @@ public enum MappingFormat {
 		}
 
 		return this.reader.read(path, progressListener);
-	}
-
-	@Nullable
-	public MappingsWriter getWriter() {
-		return this.writer;
-	}
-
-	@Nullable
-	public MappingsReader getReader() {
-		return this.reader;
-	}
-
-	public FileType getFileType() {
-		return this.fileType;
-	}
-
-	/**
-	 * Determines the mapping format of the provided file. Checks all formats according to their {@link #getFileType()} file extensions.
-	 * If the file is a directory, it will check the first file in the directory.
-	 * Will return {@link #PROGUARD} if no format is found for single files, and {@link #ENIGMA_DIRECTORY} if no format is found for directories.
-	 * @param file the file to analyse
-	 * @return the mapping format of the file.
-	 */
-	public static MappingFormat parseFromFile(Path file) {
-		if (Files.isDirectory(file)) {
-			try {
-				File firstFile = Arrays.stream(Objects.requireNonNull(file.toFile().listFiles())).findFirst().orElseThrow();
-
-				for (MappingFormat format : values()) {
-					if (!format.getFileType().isDirectory()) {
-						continue;
-					}
-
-					String extension = MoreFiles.getFileExtension(firstFile.toPath()).toLowerCase();
-					if (format.fileType.getExtensions().contains(extension)) {
-						return format;
-					}
-				}
-
-				return ENIGMA_DIRECTORY;
-			} catch (Exception e) {
-				return ENIGMA_DIRECTORY;
-			}
-		} else {
-			String extension = MoreFiles.getFileExtension(file).toLowerCase();
-
-			for (MappingFormat format : values()) {
-				if (format.fileType.getExtensions().contains(extension)) {
-					return format;
-				}
-			}
-		}
-
-		return PROGUARD;
 	}
 }
