+++ enigma/src/main/resources/META-INF/services/org.quiltmc.enigma.api.EnigmaPlugin
@@ -1,1 +1,1 @@
-org.quiltmc.enigma.impl.analysis.BuiltinPlugin
+org.quiltmc.enigma.impl.plugin.BuiltinPlugin
